requirements.txt
asgiref==3.7.2
Django==5.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
shortuuid==1.0.11
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
